requirements.txt
pathlib
openai
SpeechRecognition
pydub
python-dotenv